Network rearchitecture. (#3)
The network was rearchitected so that the app containers will now run on their own, but isolated to the workbench network. This is needed to allow for multiple NIMs to run on the same project.

Upgrading to this version:

Within the workbench GUI, synchronize with the server
Remove your /code/config.yaml file to have a new one generated.
Changes:

LLM NIM has been renamed to LLM-NIM-0
another LLM-NIM-1 was added
LLM NIM, Milvus, and Redis have been migrated to the workbench network.
app containers no longer have the project name prefix
env variables have been added to configure the networked milvus and redis.
default config file now points to the networked nim-0